refactor: remove utils.ts
